requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
aiogram
sqladmin
qrcode
itsdangerous
redis
pillow
alembic
